Use /publicsign instead of FakeSign
The new toolset supports the /publicsign command line option.  This removes the need to use FakeSign in our builds.

This commit simply removes the use of FakeSign in our builds.  It does not remove the project because that would break the internal build.  Once this commit is merged I will remove the reference to the FakeSign project in the internal build, then the public one.  That will avoid a cross cutting breaking change over the holidays.