Fixed the WPF builds
The WPF projects now need to explicitly set the `$(LanguageTargets)`
property. In the past when we controlled the manual invocation of the
core targets we could rely on using the `$(ProjectLanguage)` property to
calculate this in Imports.targets.

That is no longer possible given that the SDK controls the time in
which core targets are imported. That occurs in a place where it's not
possible for us to use `$(ProjectLanguage)` for the tmp_proj file.