Initial .NET Framework ARM64 support
This adds the initial support for Roslyn producing .NET Framework ARM64
binaries. This is necessary because on an ARM64 machine a binary tagged
as AnyCPU will actually run under X64 emulation. This meas the Roslyn
tools will run out of the box but they do so under emulation which is
significantly slower. In order to get the performance expected the
binaries need to be targetted to ARM64 so they run natively.

There are a few challenges here. The first is that there is not a
concept of multi-targeting between `<PlatformTarget>` values. In order
to support producing both AnyCPU and ARM64 EXEs there are two approaches
we could take:

1. Custom build logic to hack in the ability to multi-target on
   `<PlatformTarget>`.
2. Have a build specific for .NET Framework ARM64.

I spent several days trying to find a good path for (1) but wasn't happy
with any of my results. (2) is comparatively simple though and this PR
goes that route with the goal of making progress on this effort. The
MSBuild team is continuing to work on ways to make (1) easier and it's
my hope that we'll be able to go that route in the next few weeks. For
now though (2) is the approach because it allows forward progress.

The second challenge is there are no CI machines with .NET Framework
ARM64 installed on them. The only way to run tests is to do so manually
on a machine. There are 3 test failures when I did this but they are
pretty minor issues. Nothing that would block us getting the binaries
out for the rest of the effort to move forward.

The overall effort here and remaining work items are tracked here

https://github.com/dotnet/roslyn/issues/59421